+++ NEWS.md
@@ -178,6 +178,9 @@ Standard library changes
 
 #### InteractiveUtils
 
+* New macro `@trace_compile` for running an expression and show which methods are compiled, with timing
+  i.e. with `--trace-compile=stderr --trace-compile-timing` enabled for that call only. ([#55915])
+
 Deprecated or removed
 ---------------------